WIP: Add new entries to completion dialog
This patch adds a new version of MethodReferenceCompletionProvider that
uses FunctionalInterfaceSuggester#suggestFunctionalInterfaces which
returns the list of functionalinterfaces that a method reference can be
casted to. The list of functionalinterfaces is provided into the
completion dialog.

Known issues:

1. The element pattern does not work, because when the completion dialog
is invoked the AST is in the broken state and there is no rigig method
refenence structure, which makes impossible to provide an obvious
pattern. A possible solution: provide my own parser and lexer in order
to maintain the keep the valid PSI;
2. FunctionalInterfaceSuggester provides just a list of functional
interfaces without substituting the information about Generic types. The
straightforward solution is too primitive, intellij has to provide a
means to easily substitute these generic types;
3. The actual cast in does not occur in the editor yet;
4. The code is ugly.

Signed-off-by: Nikita Eshkeev <kastolom@gmail.com>